Prevent certain text strings from being allowed as asset names
We want to avoid overly-broad search-and-replace havoc that
certain asset names can cause.

For example, if the asset detection sees

"function url(link)"

in a file, and decides to treat "link" as an asset name, don't
permit this, even if the page being cached serves up a 200 response
when attempting to fetch it.